+++ pkg/plugin/bodylmt/setup_test.go
@@ -1,0 +1,17 @@
+package bodylmt
+
+import (
+	"testing"
+
+	"github.com/hellofresh/janus/pkg/plugin"
+	"github.com/hellofresh/janus/pkg/proxy"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestSetup(t *testing.T) {
+	route := proxy.NewRoute(&proxy.Definition{})
+	err := setupBodyLimit(route, plugin.Params{})
+	assert.NoError(t, err)
+
+	assert.Len(t, route.Inbound, 1)
+}